requirements.txt
asgiref==3.5.2
celery==5.2.7
Django==4.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
redis==4.3.4
google-api-python-client==2.58.0
